requirements.txt
# Core dependencies
flask==3.0.0
flask-cors>=4.0.0
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google Gemini AI
google-generativeai>=0.8.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0
colorama==0.4.6

# Security
cryptography>=42.0.0

# HTTP client with better error handling
httpx==0.25.2

# JSON schema validation
jsonschema==4.20.0

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0
